requirements.txt
chromadb==0.5.23
crewai==0.102.0
crewai_tools==0.36.0
gTTS==2.5.4
langchain==0.3.18
omegaconf==2.3.0
pandas==2.2.3
Pillow==11.1.0
SpeechRecognition==3.12.0
streamlit==1.40.1